Nothing in this repository had ever run clippy — not CI, not `cargo
test`, not the build (`issues/build/clippy-has-never-run-here.md`). This
is stage 1 of the owner's order to put it in CI for every pull request:
**default clippy, warnings denied, over three trees, with every finding
fixed at the site.**

Stage 2 — `pedantic` and `nursery`, one lint at a time with each one's
count measured first — is deliberately not here. Neither is
`undocumented_unsafe_blocks`, which belongs to the unsafe track.

## The gate

A `clippy` step in `host-tests.yml`'s `host` job:

```
cargo clippy --workspace --all-targets --keep-going -- -D warnings
(cd kernel     && cargo clippy --target x86_64-unknown-none -- -D warnings)
(cd bootloader && cargo clippy --target x86_64-unknown-uefi -- -D warnings)
```

**A step of `host` rather than a job of its own, and that is what makes
it bind.** `landing.yml`'s `gate-stage` reads main's required checks
back out of GitHub and reds on a name this repository claims that the
ruleset does not require — so a new job would be a check nobody has
required yet: visible on the pull request and stepped straight past by
`gh pr merge --auto`. Inside `host`, which is required today, a finding
blocks a merge today. Splitting it out later is a two-line move plus a
line in `gate-stage`'s list.

All three run on **stable**, and none waits for the ToyOS toolchain:
`x86_64-unknown-none` and `x86_64-unknown-uefi` are built-in targets.
The reason usually given for skipping the kernel — that the `toyos`
toolchain ships no clippy — was about the wrong toolchain.

**The bootloader answer: yes, it lints**, with one qualification.
`--all-targets` fails there and on the kernel, because a test target
links `std` whose `panic_impl` collides with the one `uefi_services`
provides. So `--all-targets` is the workspace's alone, where it is what
reaches `tests/`, `examples/` and every crate's test targets.

**Userland is out of scope and says so in the step**:
`x86_64-unknown-toyos` is a custom target, so it needs the fork's clippy
and the bootstrapped sysroot this job does not wait for.

Measured on the dev host, fresh target directories, warm registry:
**19.2 s** workspace (139 crates), **2.2 s** kernel (19), **3.7 s**
bootloader (20) — 25 s against this job's 398 s cold. The cache now
keeps `kernel/target` and `bootloader/target`.

## The findings

~190, all fixed at the site. Every edit is either the standard-library
spelling of what the code already computed, or an `allow` carrying its
reason at the site. **No blanket allows and no crate-level allows.**

`--keep-going` is what made the real count knowable. A crate that fails
stops every crate behind it, so the first measurement of this workspace
measured a prefix: `toyos-ld` failed (it already declared `[lints.rust]
warnings = "deny"`) and hid its own bin, `toyos-cc` (26),
`toyos-build`'s lib and bin, `examples/imgstat.rs` and the entire QEMU
harness in `tests/`.

### Three real defects, not style

- **`mut_from_ref`** — `io_uring`'s `cqe_at_mut` minted a `&mut
IoUringCqe` from `&self`, over a page the owning process also maps; two
callers could have held one at once and the compiler was told they could
not. Now `cqe_at` returning `*mut`, written through once under the lock
the caller already holds.
- **`only_used_in_recursion`, twice** — `src/sourcegate.rs`'s
`rust_files` and `src/hostws.rs`'s `build_scripts` each took a `root`
they passed down and never read. Deleted, not underscored.
- **`unnecessary_literal_unwrap`** — `toyos-cc`'s `merge_ty` was
`Some(...)` two lines above its `unwrap_or_else`, so the fallback was
unreachable.

### The two `named constant with interior mutability` verdicts: **not
bugs**

`kernel/src/log/shard.rs`'s `EMPTY` and
`kernel/src/iommu/vtd/fault.rs`'s `FaultUnit::EMPTY` each have exactly
one use — an array repeat that builds N independently-zeroed elements —
and neither is ever borrowed or written. Every store in the IOMMU file
names `UNITS[index]`, which is the `static`. The decisive check:
clippy's companion lint **`borrow_interior_mutable_const`, the one that
fires on the shape where a write is lost, is silent on both**. `const`
stays; the reason is at each site.

### The xHCI precedence findings: already `(a + b) | c`

Which is what the spec wants — CRCR bit 0 is RCS over a 64-byte-aligned
pointer (xHCI 1.2 §5.4.5), and a TRB address is 16-byte aligned with the
cycle bit in bit 0. The parentheses move no bit; they mean a reader no
longer needs the precedence table.

### Where clippy is wrong, the allow says so

The one to read is `emit_elf.rs`'s `needless_range_loop` over
`dynsym_count`: **the suggestion would be a defect.** `dynsym_count`
counts exports and shared imports too while `import_str_ids` is indexed
only on one arm, so iterating that vector instead ends the loop early
for a shared object — a different file, silently.

The rest: a loop variable that is a CPU number rather than a walk (5),
an offset table written as `N * PAGE` or `PORT + N` so it reads as one
column (3), a spec's own name (`IMAGE_SUBSYSTEM_EFI_*`, GCC's
`StmtExpr`), a `match` whose empty arm holds the argument for why it may
be empty, a `drop` marking where the kernel's bracket closes, and three
`too_many_arguments` where the arguments are a wire format's fields.

### `toyos-elf`'s 25 `dead_code`

Judged one at a time, not silenced: `Elf::poke` is used by nothing and
is deleted; the other 24 are two test binaries each driving half of one
shared builder, declared at the two `mod common;` sites rather than
inside the module.

## Filed rather than fixed

- `issues/kernel/futex-wait-never-returns-its-timeout-code.md` — found
while clearing `if_same_then_else`: `process::futex_wait` documents `1`
on timeout and no path can produce it.
- `issues/build/clippy-has-no-local-half.md` — clippy reds in CI and
nowhere before it. A pre-push hook is what would fit; this repository
has no hook mechanism, and that is a decision to make once for hooks
rather than to smuggle in under clippy.
- A sighting into the index that owns it: `xhci_full_speed_device` red
in this branch's first full suite, `ALONE … GREEN`, on a parallel phase
of 103.9 s that took 32.0 s warm minutes later.

## Verification

`cargo test` 265 passed. `cargo test --lib` 137 passed. `cargo test
--workspace --exclude toyos-build` green across every member. All three
clippy invocations exit 0.

Depends on #131, which carries the four `toyos-abi` findings —
`toyos-abi/src` compiles into the shared sysroot, so `--abi-split-check`
requires it to land alone and first.

+++ kernel/src/loader/start.rs
@@ -48,7 +48,7 @@ pub(crate) fn alloc_kernel_stack(
         *frame.add(4) = 0; // rbx
         *frame.add(5) = 0; // rbp
         *frame.add(6) = 0x002; // RFLAGS (IF=0, AC=0)
-        *frame.add(7) = trampoline as u64; // return address
+        *frame.add(7) = trampoline as usize as u64; // return address
     }
     Some((alloc, frame as u64))
 }
@@ -353,7 +353,7 @@ pub fn build_child_handles(
     drop(data);
     drop(displaced);
 
-    let mut raw = alloc::vec![0u8; endow.len() as usize];
+    let mut raw = alloc::vec![0u8; endow.len()];
     endow.read_at(0, &mut raw);
     Ok(PendingHandles::Moving { table: handles, endow: raw, labels: labels.to_vec() })
 }
